PR-SUG: move API controller to different file
